TRIV/requirements.txt
aiokafka==0.12.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels_redis==4.2.1
charset-normalizer==3.4.0
confluent-kafka==2.6.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==4.2.16
django-appconf==1.0.6
django-cors-headers==4.6.0
django-staticfiles==1.2.1
djangorestframework==3.15.2
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kafka-python-ng==2.2.3
msgpack==1.1.0
packaging==24.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.3.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.1.1
